% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sites.R
\name{combine_sites}
\alias{combine_sites}
\title{combine_sites}
\usage{
combine_sites(N, file.names, folder, site.coords, random = FALSE)
}
\arguments{
\item{N}{Integer. Number of sites to combine from folder}

\item{file.names}{Array >=1. Character strings of file names in folder}

\item{folder}{Folder than contains files}

\item{site.coords}{Data frame. Contains unique ID for sites and coordinates}

\item{random}{Logical. Randomly select N files in folder?}
}
\description{
Combines climate data from sites from folder
}
\examples{
\dontrun{
set.seed(876)
data <- combine_sites(N=1,file.names,folder='data/site_data',site.coords,random=T)
}
}
