% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_pls.R
\name{daily_pls}
\alias{daily_pls}
\title{daily_pls}
\usage{
daily_pls(data, lag = 365, ncomps = 3)
}
\arguments{
\item{data}{Data frame that consists of columns named date, month, day, and the time series
to be used for pls regression. In version 0.0.1, the response data must be called "cfs", and the
explanatory time series must be called "p", "tmax", and "tmin".}

\item{lag}{Number of lags to be used for creating partial least squares components}

\item{ncomps}{Number of components for pls regression}
}
\description{
Performs pls regression on daily data for a given number of lags.
}
\examples{
\dontrun{
data <- data
daily_pls(data,lag=20,ncomps=2)
}
}
