% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_def.R
\docType{methods}
\name{plot,timepls_fit-method}
\alias{plot,timepls_fit-method}
\title{Plots timepls_fit S4 object's date and residuals}
\usage{
\S4method{plot}{timepls_fit}(x, y, ...)
}
\arguments{
\item{x}{time_pls fit object from time_pls function.}

\item{y}{not used}
}
\description{
Plots timepls_fit S4 object's date and residuals
}
