% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_plot.R
\name{plot_cor}
\alias{plot_cor}
\title{plot_cor}
\usage{
plot_cor(fit, window, smooth)
}
\arguments{
\item{fit}{time_pls fit object from time_pls function}

\item{window}{window for correlations}

\item{smooth}{smoothing window for correlation plot}
}
\description{
Plots rolling correlation for timepls_fit S4 object's date and residuals
}
\examples{
\dontrun{
d <- climate_data
dates <- d$date
y <- d$cfs
X <- cbind(d$p,d$tmin,d$tmax)
fit <- time_pls(y,X,dates,lag=30,ncomps=10)
plot_cor(fit,window=60,smooth=90)
}
}
