% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_pls.R
\name{time_pls}
\alias{time_pls}
\title{time_pls}
\usage{
time_pls(y, X, dates, lag = 30, ncomps = 3)
}
\arguments{
\item{y}{Vector of variable to predict using pls}

\item{X}{Matrix of covariates for pls}

\item{lag}{Number of lags to be used for creating pls components}

\item{ncomps}{Number of components for pls regression}
}
\description{
Performs pls regression on time series data for a given number of lags.
}
\examples{
\dontrun{
data <- data
dates <- data$date
y <- data$cfs
X <- cbind(data$p,data$tmin,data$tmax)
plsm1 <- time_pls(y,X,dates,lag=30,ncomps=3)
}
}
