% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_def.R
\docType{class}
\name{timepls_fit-class}
\alias{timepls_fit-class}
\alias{timepls_fit}
\title{An S4 class to return the results of time_pls
\code{\link{time_pls}}.}
\description{
An S4 class to return the results of time_pls
\code{\link{time_pls}}.
}
\section{Slots}{

\describe{
\item{\code{residulas}}{residuals from fit model}

\item{\code{dates}}{dates of time series}
}}

