% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_def.R
\docType{class}
\name{timepls_fit-class}
\alias{timepls_fit-class}
\alias{timepls_fit}
\title{An S4 class to return the results of time_pls}
\description{
An S4 class to return the results of time_pls
}
\section{Slots}{

\describe{
\item{\code{observed}}{observed values}

\item{\code{estimated}}{estimated values}

\item{\code{residuals}}{residuals from fit model}

\item{\code{dates}}{dates of time series}
}}

